fix: Default maximize_build_space to true
This comes as a realization that we've had many several users run into this issue only to be told they need to change an input for the action. I feel that defaulting this to true is a saner default if only to allow users to build without eventually running into this problem as they increase their image size. It seems to confuse more than help with default false.