fix: use github runner token when calling the api, to prevent rate li… (#124)
* fix: prevent rate limiting by using github token

Sometimes the runner can get rate-limited during the step where it
calls the API to determine the latest tag. This fixes that by passing
the runner's github token as an authentication token.

Using curl's built-in variable support, we can avoid passing the
plaintext secret on the command line as an argument, thus avoiding the
theoretical security issue of someone being able to view the command
line args using `ps`.

* oh shit i thought i pushed this, my bad